TEST: To use new git template
Problem
To create a standardized way to have a more detailed and verbose commit
strategy. Jeremy Duckworth sent me this initiative and I would like to
think I'm nailing it. The commit is just to test the new structure.

Solution
Create a new template in ~/.gitmessage and was able to use the command
> git config --global commit.template ~/.gitmessage
in order to make this my automatic template for my local machine.

Note
* This can be overriden with a normal
> git commit -m <your msg here>